Breaking a part main RealTimeDIC file into Classes and Widgets modules
